backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI Providers
openai==1.12.0
anthropic==0.18.0
httpx==0.26.0

# Vector DB
qdrant-client==1.7.0

# Billing
stripe==8.0.0

# Document processing
pypdf==4.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
